Refactorizacion de persitencia
