--- SistemaContable/src/Persistencia/Lector.java
@@ -1,20 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Persistencia;
-/**
- *
- * @author pc
- */
-public class Lector {
-    private String ruta;
-  
-
-    public Lector(String ruta) {
-        this.ruta = ruta;
-       
-    }
-    
- 
-}
